Bound oversized review and commit prompts
- Share bounded diff and context summarization across review and commit-message generation.
- Retry input-size failures with progressively smaller prompts within a fixed shared provider-call budget, including repairs and transport retries.
- Disclose limited review coverage when summaries omit original detail.
