+++ crates/ag-agent/src/lib.rs
@@ -12,7 +12,9 @@ mod agent;
 mod app_server;
 pub(crate) mod app_server_transport;
 mod channel;
+mod input_size;
 mod model;
+mod provider_call_budget;
 
 pub use agent::{
     AgentAvailabilityProbe, AgentBackend, AgentBackendError, AgentTransport, BuildCommandRequest,
@@ -36,6 +38,7 @@ pub use channel::{
 };
 #[cfg(any(test, feature = "test-utils"))]
 pub use channel::{MockAgentChannel, create_cli_agent_channel_with_backend};
+pub use input_size::is_input_size_error;
 pub use model::agent::{
     AgentCliInfo, AgentCliVersion, AgentKind, AgentModel, AgentSelection, AgentSelectionMetadata,
     ReasoningLevel, parse_persisted_session_agent_model, resolve_agent_kind_for_model,
@@ -44,3 +47,4 @@ pub use model::agent::{
 };
 pub use model::permission::PermissionMode;
 pub use model::session::{ResponseStyle, SessionDiffState, SessionStats, SpeedMode};
+pub use provider_call_budget::ProviderCallBudget;
